requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.4
antlr4-python3-runtime==4.9.3
asgiref==3.11.1
asteroid-filterbanks==0.4.0
asttokens==3.0.1
attrs==25.4.0
av==17.0.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
Django==6.0.2
django-debug-toolbar==6.2.0
django-pwa==2.0.1
django-unfold==0.85.0
django-webpush==0.3.6
einops==0.8.2
executing==2.2.1
filelock==3.24.3
flatbuffers==25.12.19
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
googleapis-common-protos==1.73.1
greenlet==3.3.2
grpcio==1.80.0
gunicorn==25.3.0
hf-xet==1.4.2
http_ece==1.2.1
huggingface_hub==0.36.2
idna==3.11
importlib_metadata==8.7.1
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
julius==0.2.7
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.5.0
lightning-utilities==0.15.3
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.1
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.4
numpy==2.4.2
packaging==26.0
pandas==3.0.1
parso==0.8.6
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.2
playwright==1.58.0
primePy==1.3
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.6
psutil==7.2.2
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
py-vapid==1.9.4
pycparser==3.0
pyee==13.0.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pywebpush==2.3.0
PyYAML==6.0.3
pyzmq==27.1.0
regex==2026.2.19
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==82.0.0
six==1.17.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.48
sqlparse==0.5.5
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
transformers==4.57.6
typing_extensions==4.15.0
urllib3==2.6.3

# Ajouter ces deux lignes AVANT openai-whisper
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.7.0+cpu
torchaudio==2.7.0+cpu
openai-whisper==20250625
wcwidth==0.6.0
whitenoise==6.12.0
yarl==1.23.0
zipp==3.23.0
